Prefer to call `update_state` directly
Instead of message passing by returning the future-like return that
`message::command:message` produces.

This `command_message` approach causes none of these messages to be
emitted until the app has focus (for me in sway, at least), so we're
just calling the function directly, to get the effect we want, right
away.