Copy to buildexec directory as part of deploy
This "project" is a syncthing target for me, where it holds just the
compiled binaries I need to run this project. It's a bit hacky, but it's
working for now. The real end game is to have this packaged up as
something I can install with pacman or something, but I'm not at that
scale yet, I claim.

Note that this other "project" does have its own `install.sh` script
that is a copy-paste of the %USERNAME% sed, which I'm just admitting as
a DRY-fail for the moment.