Move parser logic to separate module
We do make the display_debug method public as part of this; that could
have been in its own commit, sorry.

We also #[allow(unused)] on the `simple_debug_display` function, as it
was allowed to be public from the top level module of this crate,
because it was public.