Initial youtube importer logic
This is just a chunk of what I have working, if you're ever reading
this: I'm sorry, I should have broken this up more. It doesn't even make
sense to stop here, as all that we do is create the manual artist
mapping. Will keep adding to do the track mapping, and actually produce
the count file that the actual app reads.