Initial inversion of backend controlling playback
Instead of the app relaying mpris and sink actions to each other (and
the tracker once we finish this) instead this backend layer controls
this instead. This doesn't actually do everything we need, we don't set
the mpris data as we don't accept this information in our cheaty
sink-only message from the gui and we don't set the mpris to stopped
once we reach the end, and especially of note we don't actually control
the play queue from this new backend. This is one huge messy commit, if
you read this and are trying to dissect it: I'm sorry, I got lazy here.