Intial `preferred_region` logic
This allows us to try to find the correct box art based on the region
the user specifies in their config.

Note that we need to rework how name/path maps are stored as right now
we only have one name to match against in our best_match logic. That's
next, stay tuned!