Simplify `find_files` type signature
Instead of requiring the same path twice, introduce a helper function
that takes both paths and expose a function that just accepts the one
path to operate over. This will also be useful as we try to `threadpool`
this function and have accept a `tx` that we won't want exposed on the
public function either.