Use `only_files` return in `full_file` lookup
We are going to change this for loop to use a threadpool and it will be
much easier to do so over a for loop, rather than re-implementing the
recursive lookup. This helps with some repetition we were doing with
these similar recursive functions, now we only have one canonical
recursive file-finding function (the `find_only_files` function, which
was also confirmed to be very fast with the `musiqinspector` mode added
in the last commit).