Don't pass/accept/store sink volume outside of sink
This was being incorrectly reset to 1.0 on each new track, as the second
copy of the sink volume was never set from the layer below it. This
removes the need for that assignment entirely and just has the sink
re-use the volume from it's last track, which should always be correct.